Okta app sync/search .. split out common methods for sync and search and implement an abstract class for the API client
