+++ src/Services/OktaAppUserSync.php
@@ -12,50 +12,10 @@ use SilverStripe\Security\Member;
 use SilverStripe\Security\Permission;
 
 /**
- * Synchronises the users assigned with the application
+ * Synchronises the users assigned to the application with the local Member records
  */
-class OktaAppUserSync
+class OktaAppUserSync extends OktaAppClient
 {
-
-    use OktaGroups;
-
-    use Configurable;
-
-    /**
-     * A user assigned to an application directly
-     */
-    const APPUSER_SCOPE_USER = 'USER';
-
-    /**
-     * A user assigned to an application via a group
-     */
-    const APPUSER_SCOPE_GROUP = 'GROUP';
-
-    /**
-     * @var \Okta\Client|null
-     */
-    private $client = null;
-
-
-    /**
-     * @var bool
-     */
-    private $dryRun = false;
-
-    /**
-     * @var string
-     */
-    private $start = '';
-
-    /**
-     * @var array
-     */
-    private $report = [];
-
-    /**
-     * @var HttpClient
-     */
-    protected $httpClient = null;
 
     /**
      * @var bool
@@ -65,62 +25,14 @@ class OktaAppUserSync
     private static $create_users = true;
 
     /**
-     * Get the configured {@link \Okta\Client}, if not available create it from configuration
-     */
-    protected function getClient($parameters = []) : \Okta\Client
-    {
-        if (!$this->client) {
-            $this->httpClient = new ProxiedCurlHttpClient();
-            $this->client = ClientFactory::create($parameters, $this->httpClient);
-        }
-        return $this->client;
-    }
-
-    /**
-     * Client ID for the application being synchronised
-     */
-    protected function getClientId() : string
-    {
-        return Config::inst()->get(ClientFactory::class, 'application_client_id');
-    }
-
-    /**
-     * Return the sync report, which is an array, keys are Okta User Ids, each value is an array
-     * of changes performed on that user
-     * Report is only gathered in dryRun mode
-     * @return array
-     */
-    public function getReport() : array
-    {
-        return $this->report;
-    }
-
-    /**
-     * Return the users successfully sync as an array, keys are Okta user ids, values is the match Member.ID
-     * @return array
-     */
-    public function getSuccesses() : array
-    {
-        return $this->success;
-    }
-
-    /**
-     * Return the failed sync attempts, array of okta user id values
-     * @return array
-     */
-    public function getFailures() : array
-    {
-        return $this->fail;
-    }
-
-    /**
      * Run the sync processing
      * @see https://developer.okta.com/docs/reference/api/apps/#list-users-assigned-to-application
      * @param bool $dryRun when true, no changes are made to local members. A report is created and can be accessed via getReport()
      * @param int $limit the number of records to return per page, if 0 the default Okta value is used (50 based on docs)
+     * @param array $queryOptions extra filtering options to pass to the Okta API
      * @throws \Exception
      */
-    public function run(bool $dryRun = false, int $limit = 50) : int
+    public function run(bool $dryRun = false, int $limit = 50, array $queryOptions = []) : int
     {
         $this->success = $this->fail = [];
         $this->dryRun = $dryRun;
@@ -131,7 +43,7 @@ class OktaAppUserSync
         // create/configure the Okta client
         $client = $this->getClient();
 
-        $this->getAppUsers($limit);
+        $this->getAppUsers($limit, $queryOptions);
 
         $dt = new \DateTime();
         $this->start = $dt->format('Y-m-d H:i:s');
@@ -139,61 +51,6 @@ class OktaAppUserSync
         $failCount = count($this->fail);
         Logger::log("OKTA: processUsers complete, {$successCount} users successfully synced, {$failCount} fails", "INFO");
         return $successCount;
-    }
-
-    /**
-     * Collect all app users via pagination method
-     * @param
-     * @return void
-     */
-    private function getAppUsers(int $limit = 50)
-    {
-
-        // Initial set
-        $this->appUsers = new \Okta\Applications\Collection([]);
-
-        // initial properties for App users request
-        $properties = new \stdClass;
-        $properties->id = $this->getClientId();
-        if (empty($properties->id)) {
-            throw new \Exception("No App ClientId configured (ClientFactory.application_client_id)");
-        }
-        $resource = new \Okta\Applications\Application(null, $properties);
-
-        // initial options for initial request
-        $options = [
-            'query' => [
-                'limit' => $limit
-            ]
-        ];
-        $this->collectAppUsers($options, $resource);
-    }
-
-    /**
-     * Get all appusers based on configuration
-     * @param array $options
-     * @param Okta\Applications\Application $resource
-     */
-    private function collectAppUsers(array $options, \Okta\Applications\Application $resource)
-    {
-        // @var \Okta\Applications\Collection
-        $collection = $resource->getApplicationUsers($options);
-        if ($collection instanceof \Okta\Applications\Collection) {
-            // merge the returned collection on
-            if (!$this->appUsers) {
-                $this->appUsers = $collection;
-            } else {
-                $this->appUsers = $this->appUsers->merge($collection);
-            }
-            try {
-                $options = $this->httpClient->getNextPageOptions();
-                // get the next set
-                $this->collectAppUsers(['query' => $options ], $resource);
-            } catch (\Exception $e) {
-                // getNextPageOptions threw an exception (or no more results)
-            }
-        }
-        return false;
     }
 
     /**
@@ -223,33 +80,6 @@ class OktaAppUserSync
             }
         }
         return count($this->success);
-    }
-
-    /**
-     * Collect all groups for a user
-     * @param array $options
-     * @param \Okta\Users\User $resource
-     * @param \Okta\Groups\Collection $userGroups
-     */
-    private function collectUserGroups(array $options, \Okta\Users\User $resource, \Okta\Groups\Collection &$userGroups)
-    {
-        // @var \Okta\Groups\Collection
-        if (!$resource->getId()) {
-            throw new \Exception("To get user groups, the user resource must have an Id");
-        }
-        $collection = $resource->getGroups($options);
-        if ($collection instanceof \Okta\Groups\Collection) {
-            // merge the returned collection on
-            $userGroups = $userGroups->merge($collection);
-            try {
-                $options = $this->httpClient->getNextPageOptions();
-                // get the next set
-                $this->collectUserGroups(['query' => $options ], $resource, $userGroups);
-            } catch (\Exception $e) {
-                // getNextPageOptions threw an exception (or no more results)
-            }
-        }
-        return false;
     }
 
     /**
@@ -293,12 +123,12 @@ class OktaAppUserSync
      */
     protected function processAppUser(\Okta\Applications\AppUser $appUser) : Member
     {
-        $userId = $appUser->getId();
 
-        // Get the corresponding user record
-        $userResource = new \Okta\Users\User();
-        $user = $userResource->get($userId);
+        // Get the user record, incl Okta userId
+        $user = $this->getUser($appUser);
+        $userId = $user->getId();
 
+        // Collect user groups
         // initial request options
         $options = [
             'query' => [
@@ -309,15 +139,12 @@ class OktaAppUserSync
         $userGroups = new \Okta\Groups\Collection([]);
         $this->collectUserGroups($options, $user, $userGroups);
 
-        // @var \Okta\Users\UserProfile
-        $userProfile = $user->getProfile();
-        if (!($userProfile instanceof \Okta\Users\UserProfile)) {
-            throw new OktaAppUserSyncException("AppUser {$userId} has no Okta user profile");
-        }
+        // User profile information
+        $userProfile = $this->getUserProfile($user);
 
         // @var string - either USER or GROUP
         $appUserScope = $appUser->getScope();
-        Logger::log("AppUser.id={$userId} User.id={$user->getId()} scope={$appUserScope}", "DEBUG");
+        Logger::log("AppUser.id={$appUser->getId()} User.id={$user->getId()} scope={$appUserScope}", "DEBUG");
 
         // @var string
         $userUsername = $userProfile->getLogin();
@@ -406,18 +233,6 @@ class OktaAppUserSync
         }
 
         return $member;
-    }
-
-    /**
-     * Ensure that a profile has HTML removed
-     */
-    private function sanitiseProfileValue($value) : string
-    {
-        if (is_scalar($value)) {
-            return trim(strip_tags($value));
-        } else {
-            return '';
-        }
     }
 
     /**
